WIP: Continue Graph class update
Use multiple maps (vertexes, reverseVertexes, edges, reverseEdges)

Faced with problem:
reverseEdges: Map<{from: number, to: number}, EdgeId>
If we create a new object {from: ..., to: ...} and try to use as a key to
get value from the map, we'll fail. It fails because JS compares objects
as references.

So current implementation doesn't work.